Chore: fmt, vm labels, feat:NovaFactoryI

+++ contracts/IAutID.sol
@@ -10,11 +10,7 @@ interface IAutID is IERC721Upgradeable {
     event AutIDCreated(address owner, uint256 tokenID);
     event DAOJoined(address daoExpanderAddress, address member);
     event DAOWithdrown(address daoExpanderAddress, address member);
-    event CommitmentUpdated(
-        address daoExpanderAddress,
-        address member,
-        uint256 newCommitment
-    );
+    event CommitmentUpdated(address daoExpanderAddress, address member, uint256 newCommitment);
     event DiscordIDConnectedToAutID();
     event MetadataUriSet(uint256 tokenId, string metadataUri);
 
@@ -31,76 +27,49 @@ interface IAutID is IERC721Upgradeable {
     /// @param role the role that the user has selected within the specified DAO
     /// @param commitment the commitment value that the user has selected for this DAO
     /// @param daoExpander the address of the daoExpander contract
-    function mint(
-        string memory username,
-        string memory url,
-        uint256 role,
-        uint256 commitment,
-        address daoExpander
-    ) external;
+    function mint(string memory username, string memory url, uint256 role, uint256 commitment, address daoExpander)
+        external;
 
     /// @notice associates a AutID to a new DAO
     /// @dev The commitment of the user can't exceed 10. It fails if the user has already committed to other DAOs
     /// @param role the role that the user has selected within the specified DAO
     /// @param commitment the commitment value that the user has selected for this DAO
     /// @param daoExpander the address of the daoExpander contract
-    function joinDAO(
-        uint256 role,
-        uint256 commitment,
-        address daoExpander
-    ) external;
+    function joinDAO(uint256 role, uint256 commitment, address daoExpander) external;
 
     /// @notice gets all communities the AutID holder is a member of
     /// @param autIDHolder the address of the AutID holder
     /// @return daos dao expander addresses that the aut holder is a part of
-    function getHolderDAOs(address autIDHolder)
-        external
-        view
-        returns (address[] memory daos);
+    function getHolderDAOs(address autIDHolder) external view returns (address[] memory daos);
 
     /// @notice returns NFT ID of the holder
     /// @param autIDOwner the user address
     /// @return the token ID of their NFT ID
-    function getAutIDByOwner(address autIDOwner)
-        external
-        view
-        returns (uint256);
+    function getAutIDByOwner(address autIDOwner) external view returns (uint256);
 
     /// @notice AutID holders can link their DiscordID to their AutID ID
     /// @dev Returns a discord ID for the AutID token ID passed.
     /// @param tokenID the AutID token ID
     /// @return discord ID
-    function autIDToDiscordID(uint256 tokenID)
-        external
-        view
-        returns (string memory);
+    function autIDToDiscordID(uint256 tokenID) external view returns (string memory);
 
     /// @notice AutID holders can link their DiscordID to their AutID ID
     /// @dev links the discord ID to the AutID token ID passed.
     /// @param discordID the discord ID of the AutID holder
     function addDiscordIDToAutID(string calldata discordID) external;
 
-    function getMembershipData(address autIDHolder, address daoExpander)
-        external
-        view
-        returns (DAOMember memory);
+    function getMembershipData(address autIDHolder, address daoExpander) external view returns (DAOMember memory);
 
-    function getTotalCommitment(address autIDHolder)
-        external
-        view
-        returns (uint256);
+    function getTotalCommitment(address autIDHolder) external view returns (uint256);
 
     function withdraw(address daoExpander) external;
 
     function editCommitment(address daoExpander, uint256 commitment) external;
 
-    function discordIDToAddress(string calldata discordID)
-        external
-        view
-        returns (address);
+    function discordIDToAddress(string calldata discordID) external view returns (address);
 
     function setMetadataUri(string calldata metadataUri) external;
 
     function getAutIDHolderByUsername(string memory username) external view returns (address);
-    function getNextTokenID() external view returns(uint);
+    function getNextTokenID() external view returns (uint256);
 }
