+++ contracts/daoStandards/ICompound.sol
@@ -4,11 +4,10 @@ pragma solidity 0.8.18;
 /// @title ICompound
 /// @notice The minimal interface of the Compound contract needed for implementing IMembershipExtension
 interface ICompound {
-
     /// @notice Function in the Compound contract for fetching a member.
     /// @notice If for index 0 returned address is not empty, then account is a member.
     /// @param member The member address
     /// @param index Array index of the asset
     /// @return The Member struct of the address passed
-    function accountAssets(address member, uint index) external view returns(address);
+    function accountAssets(address member, uint256 index) external view returns (address);
 }
