+++ contracts/daoUtils/abstracts/DAOMarket.sol
@@ -9,15 +9,10 @@ import "../interfaces/get/IDAOMarket.sol";
 abstract contract DAOMarket is IDAOMarket {
     event MarketSet();
 
-    uint public override market;
+    uint256 public override market;
 
-    function _setMarket(uint _market)
-        internal
-    {
-         require(
-            _market > 0 && _market < 4,
-            "invalid market"
-        );
+    function _setMarket(uint256 _market) internal {
+        require(_market > 0 && _market < 4, "invalid market");
 
         market = _market;
         emit MarketSet();
