+++ contracts/daoUtils/abstracts/DAOModules.sol
@@ -10,7 +10,7 @@ import "../../plugins/registry/IPluginRegistry.sol";
 /// @dev The extension of each DAO that integrates Aut
 abstract contract DAOModules is IDAOModules {
     address public override pluginRegistry;
-    uint[] activatedModules;
+    uint256[] activatedModules;
 
     modifier onlyPluginRegistry() {
         require(msg.sender == address(pluginRegistry), "Only plugin registry");
@@ -22,22 +22,17 @@ abstract contract DAOModules is IDAOModules {
         pluginRegistry = _pluginRegistry;
     }
 
-    function _activateModule(uint moduleId) internal {
-        address modulesRegistry = IPluginRegistry(pluginRegistry)
-            .modulesRegistry();
+    function _activateModule(uint256 moduleId) internal {
+        address modulesRegistry = IPluginRegistry(pluginRegistry).modulesRegistry();
         require(
-            bytes(IModuleRegistry(modulesRegistry).getModuleById(moduleId).metadataURI)
-                .length > 0,
-            "invalid module"
+            bytes(IModuleRegistry(modulesRegistry).getModuleById(moduleId).metadataURI).length > 0, "invalid module"
         );
         activatedModules.push(moduleId);
         emit ModuleActivated(moduleId);
     }
 
-    function isModuleActivated(
-        uint moduleId
-    ) public view override returns (bool) {
-        for (uint i = 0; i < activatedModules.length; i++) {
+    function isModuleActivated(uint256 moduleId) public view override returns (bool) {
+        for (uint256 i = 0; i < activatedModules.length; i++) {
             if (activatedModules[i] == moduleId) return true;
         }
         return false;
