+++ contracts/daoUtils/interfaces/get/IDAOCommitment.sol
@@ -4,6 +4,5 @@ pragma solidity 0.8.18;
 /// @title IDAOMetadata
 /// @notice The interface for the extension of each DAO that integrates AutID
 interface IDAOCommitment {
-    function getCommitment() external view returns(uint commitment);
-
+    function getCommitment() external view returns (uint256 commitment);
 }
