+++ contracts/daoUtils/interfaces/set/IDAOCommitmentSet.sol
@@ -4,7 +4,5 @@ pragma solidity 0.8.18;
 /// @title IDAOMetadata
 /// @notice The interface for the extension of each DAO that integrates AutID
 interface IDAOCommitmentSet {
-
-    function setCommitment(uint commitment) external;
-
+    function setCommitment(uint256 commitment) external;
 }
