+++ contracts/mocks/Compound.sol
@@ -6,10 +6,9 @@ import "../daoStandards/ICompound.sol";
 /// @title Compound
 /// @notice Mock Compound DAO for testing
 contract Compound is ICompound {
-
     mapping(address => address[]) _accountAssets;
 
-    function accountAssets(address member, uint index) external view override returns(address) {
+    function accountAssets(address member, uint256 index) external view override returns (address) {
         return _accountAssets[member][index];
     }
